fix swapConfig persistency on swapWidget
